refactor: isolate emulator resize reflow
